pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dspy-cli"
version = "0.1.0"
description = "A CLI tool for creating and serving DSPy projects"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "DSPy CLI Contributors"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "dspy-ai>=3.0.4b1",
    "click>=8.0",
    "fastapi>=0.100",
    "uvicorn[standard]>=0.20",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "pydantic>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "httpx>=0.24",  # For testing FastAPI
    "black>=23.0",
    "ruff>=0.1.0",
]

[project.scripts]
dspy-cli = "dspy_cli.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 120
target-version = ["py39"]

[tool.ruff]
line-length = 120
target-version = "py39"
